dom/tests/mochitest/bugs/iframe_bug430276-2.html

<!DOCTYPE html>
<html>
<head>
  <title></title>
  <script type="application/javascript">
  new XMLHttpRequest();
  function runTest() {
    document.open();
    var succeeded = true;
    try {
      new XMLHttpRequest();
    } catch(e) {
      succeeded = false;
    }
    window.parent.ok(succeeded, "Creating XMLHttpRequest failed!");
    window.parent.SimpleTest.finish();
    document.close();
  }
  </script>
</head>
<body onload="runTest()">
</body>
</html>
